S08P12A401-20: 경로 관련 정보 화면에 띄우기

See merge request s08-webmobile4-sub2/S08P12A401!35

+++ Artwalk_Android/domain/src/main/java/com/a401/domain/model/Route.kt
@@ -1,13 +1,13 @@
 package com.a401.domain.model
 
 data class Route(
-    val duration: Double,
-    val distance: Double,
+    val duration: Int,
+    val distance: Float,
     val geometry: String,
 ) {
     constructor() : this(
-        0.0,
-        0.0,
+        0,
+        0.0f,
         ""
     )
 }
